Finally migrating default optimizers....to my own TR optimizer that uses JuMP+Ipopt to solve the local problems exactly. That's not optimal, but it means I can actually add box constraints, and hopefully also exactly solving the local problems will at least sometimes save users a few Hessian calls.
Long-term, I'd really like to find a TR/SQP solver that is gratis and
libre to make the default here. But I haven't found one that is
sufficiently satisfying to take the plunge. And on the bright side, at
least JuMP+Ipopt is rock solid, and in the long run I think using JuMP
for the outer optimizer as well will make it very easy for users to swap
in different solvers as their needs and resources vary.