Adding a `knnconfig` that has also been promoted to the default suggestion in the README and examples.
The constructor is based on `AdaptiveKDTrees.jl`, which makes me nervous
considering that that package hasn't been touched in two years and
doesn't have a lot of eyes on it. There may be edge cases where it
performs poorly. But at least for the moment, I think this is the best
option. And I am sick of not having a good simple knn-based option for
users.